Implemented "client-side prediction"
Now the clients apply the player's input into the spaceship each time it gets sent to the server, so there is no lag upon player's input application, making a "client-side prediction" simulation.

However, as the way we have programmed it all, it is pretty simple, is not a "prediction" as is, it's just the application of the input after sending it to the server, and since the server doesn't sends inputs to the clients, we didn't needed a server reconciliation since are the clients the ones making and sending the input, the server just receives it and sends the world state so all the clients update it in their machine (in fact, this replication acts kind as a server reconciliation).

Co-Authored-By: Sergi PR <t3m1X@users.noreply.github.com>